refactor: separate App into smaller components and connect them with context api
